+++ app/templates/poi.config.js
@@ -1,15 +1,10 @@
-const { name, description } = require('./package.json')
+const { name, description } = require('./package.json');
 
 module.exports = {
   html: {
     title: `${name} | ${description}`,
   },
-  entry: [
-    'storybook/config.js',
-    'storybook/addons.js'
-  ],
-  presets: [
-    require('poi-preset-storybook')()
-  ],
+  entry: ['storybook/config.js', 'storybook/addons.js'],
+  presets: [require('poi-preset-storybook')()],
   removeDist: true,
-}
+};