+++ util/pascalcase.js
@@ -1,8 +1,8 @@
-const flow = require("lodash/flow");
-const camelCase = require("lodash/camelCase");
-const upperFirst = require("lodash/upperFirst");
+const flow = require('lodash/flow');
+const camelCase = require('lodash/camelCase');
+const upperFirst = require('lodash/upperFirst');
 
 module.exports = flow(
   camelCase,
-  upperFirst
+  upperFirst,
 );