+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change Log
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
@@ -7,82 +8,123 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased][]
 
 ### Added
+
 - Adding OSS badges
+
+### Updated
+
+- Bumping dependencies
+- Bumping dev dependencies
+- Removing unused packages
+- Updating `README.md`
+- Updating package keywords
 
 ## [3.0.1][] - 2018-01-21
 
 ### Added
+
 - Adding Travis-CI integration
 
 ### Refactored
+
 - Removing `update-boilerplate` script in favour of `update-yeoman-boilerplate` repository
 
 ## [3.0.0][] - 2018-01-21
 
 ### Added
+
 - Adding Travis-CI integration
 
 ### Refactored
+
 - Removing `update-boilerplate` script in favour of `update-yeoman-boilerplate` repository
 
 ## [2.2.12][] - 2017-12-31
+
 ### Added
+
 - New ignored files/folders in `.npmignore` file
 - Bumping `update-yeoman-generator` to `1.2.1`
 
 ## [2.2.11][] - 2017-12-29
+
 ## Added
+
 - Adding `updateScripts` option to update NPM package dependencies before run `npm install`
 
 ## [2.2.10][] - 2017-12-29
+
 ## Fixes
+
 - Removing `files` information on `app/templates/package.json` to avoid npm issue when publishing package. When the package was published for some reason NPM was reading both `package.json#files` information and the publish was broken
 
 ## [2.2.9][] - 2017-12-28
+
 ## Chore
+
 - Adding MIT license
 - Updating `.gitignore` file
 
 ## [2.2.8][] - 2017-12-28
+
 ## Chore
+
 - Updating `package-lock.json` file
 
 ## [2.2.7][] - 2017-12-28
+
 ## Fixed
+
 - Adding `files` field in `package.json`
 
 ## [2.2.6][] - 2017-12-28
-## Fixed 
+
+## Fixed
+
 - Cleaning up npmignore and gitignore
 
 ## [2.2.4][] - 2017-12-28
+
 ## Fixed
+
 - Removing ignored files
 
 ## [2.2.3][] - 2017-12-28
-## Fixed 
+
+## Fixed
+
 - Writing files directly using `forEach`
 
 ## [2.2.3-0][] - 2017-12-28
+
 ## Fixed
+
 - Writing files using `for` instead of `forEach`
 
 ## [2.2.2][] - 2017-12-28
+
 ## Fixed
+
 - using `npmignore` without dot
 
 ## [2.2.1][] - 2017-12-28
+
 ## Fixed
+
 - Keep the install flow if `.nvmrc` is not available
 
 ## [2.2.0][] - 2017-12-27
+
 ## Fixed
+
 - Fixing `npm run build:page` command
 
 ## Tests
+
 - Increasing code coverage
 
 ### Added
+
 - Bumping `generator-poi-boilerplate-demo` to `1.2.0`
 - Adding tests for new template files
 - using and installing NodeJS version via NVM
@@ -91,30 +133,37 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Tests for new template files
 
 ## [2.1.1][] - 2017-12-26
+
 ### Added
+
 - Adding `update-yeoman-generator` integration
 - Removing repository prefix
 
 ## [2.1.0][] - 2017-12-25
+
 ### Added
+
 - Adding `storybook` addons
 
 ## [2.0.1][] - 2017-12-25
+
 ### Fixed
+
 - Adding `generator-poi-boilerplate-demo` component as dev dependency with how to start message
 
 ## [2.0.0][] - 2017-12-25
 
 ### Added
+
 - Adding `react-storybook` as demo wrapper
 
 ## [1.0.1][] - 2017-12-24
 
 ### Added
+
 - Created `generator-poi` project
 
-
-[Unreleased]: https://github.com/willmendesneto/generator-poi-boilerplate/compare/v3.0.1...HEAD
+[unreleased]: https://github.com/willmendesneto/generator-poi-boilerplate/compare/v3.0.1...HEAD
 [3.0.1]: https://github.com/willmendesneto/generator-poi-boilerplate/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/willmendesneto/generator-poi-boilerplate/compare/v2.2.12...v3.0.0
 [2.2.12]: https://github.com/willmendesneto/generator-poi-boilerplate/compare/v2.2.11...v2.2.12